requirements.txt
alabaster==0.7.10
alembic==0.9.8
anaconda-client==1.6.5
anaconda-navigator==1.6.9
anaconda-project==0.8.0
appnope==0.1.0
appscript==1.0.1
asn1crypto==0.22.0
astroid==1.5.3
astropy==2.0.2
Babel==2.5.0
backports-abc==0.5
backports.functools-lru-cache==1.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
backports.weakref==1.0.post1
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==1.5.0
bokeh==0.12.7
boto==2.48.0
Bottleneck==1.2.1
cdecimal==2.3
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
cloudpickle==0.4.0
clyent==1.2.2
cmake==0.9.0
colorama==0.3.9
conda==4.4.11
conda-build==3.0.22
conda-verify==2.0.0
configparser==3.5.0
contextlib2==0.5.5
cryptography==2.0.3
cycler==0.10.0
Cython==0.26.1
cytoolz==0.8.2
dask==0.15.2
datashape==0.5.4
decorator==4.1.2
distributed==1.18.3
docutils==0.14
entrypoints==0.2.3
enum34==1.1.6
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==2.0.12
Flask==0.12.2
Flask-Cors==3.0.3
flask-heroku==0.1.9
Flask-Migrate==2.1.1
Flask-Moment==0.6.0
Flask-SQLAlchemy==2.3.2
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
futures==3.1.1
gevent==1.2.2
glob2==0.5
gmpy2==2.0.8
google==1.9.3
greenlet==0.4.12
grin==1.2.1
gunicorn==19.7.1
h5py==2.7.1
heapdict==1.0.0
html5lib==0.9999999
idna==2.6
imageio==2.2.0
imagesize==0.7.1
ipaddress==1.0.18
ipykernel==4.6.1
ipython==5.4.1
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==4.2.15
itsdangerous==0.24
jdcal==1.3
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
jupyterlab==0.27.0
jupyterlab-launcher==0.4.0
Keras==2.1.2
lazy-object-proxy==1.3.1
llvmlite==0.20.0
locket==0.2.0
lxml==3.8.0
Mako==1.0.7
Markdown==2.6.10
MarkupSafe==1.0
matplotlib==2.0.2
mccabe==0.6.1
missingno==0.3.7
mistune==0.7.4
mock==2.0.0
mpmath==0.19
msgpack-python==0.4.8
multipledispatch==0.4.9
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==1.11
nltk==3.2.4
nose==1.3.7
notebook==5.0.0
np-utils==0.5.4.0
numba==0.35.0+6.gaa35fb1.dirty
numexpr==2.6.2
numpy==1.13.1
numpydoc==0.7.0
odo==0.5.1
olefile==0.44
openpyxl==2.4.8
ordereddict==1.1
packaging==16.8
pandas==0.20.3
pandas-datareader==0.5.0
pandocfilters==1.4.2
partd==0.3.8
path.py==10.3.1
pathlib==1.0.1
pathlib2==2.3.0
patsy==0.4.1
pbr==3.1.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
pipenv==11.8.0
pkginfo==1.4.1
plotly==2.0.16
ply==3.10
prompt-toolkit==1.0.15
protobuf==3.5.0.post1
psutil==5.2.2
ptyprocess==0.5.2
py==1.4.34
PyAudio==0.2.7
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0
pyflakes==1.5.0
Pygments==2.2.0
pylint==1.7.2
pyodbc==4.0.17
pyOpenSSL==17.2.0
pyparsing==2.2.0
PySocks==1.6.7
pytest==3.2.1
python-dateutil==2.6.1
python-editor==1.0.3
python-slugify==1.2.4
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.2
qgrid==0.3.3
QSTK==0.2.8
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
requests==2.18.4
requests-file==1.4.2
requests-ftp==0.3.1
rope==0.10.5
ruamel-yaml==0.11.14
scandir==1.5
scikit-image==0.13.0
scikit-learn==0.19.0
scipy==0.19.1
seaborn==0.8
simplegeneric==0.8.1
simplejson==3.11.1
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.7
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
spyder==3.2.3
SQLAlchemy==1.1.13
statsmodels==0.8.0
stevedore==1.28.0
subprocess32==3.2.7
sympy==1.1.1
tables==3.4.2
tblib==1.3.2
tensorflow==1.4.1
tensorflow-tensorboard==0.4.0rc3
terminado==0.6
testpath==0.3.1
toolz==0.8.2
tornado==4.5.2
tqdm==4.19.2
traitlets==4.3.2
typing==3.6.2
unicodecsv==0.14.1
Unidecode==1.0.22
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==2.0.1
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==0.9.8
xlwings==0.11.4
xlwt==1.2.0
yahoo-finance==1.4.0
zict==0.1.2
